requirements.txt
beautifulsoup4>=4.12.3
faiss-cpu>=1.8.0
fastapi>=0.115.0
google-genai>=1.30.0
numpy>=1.26.4
pydantic>=2.9.2
requests>=2.32.3
scikit-learn>=1.5.1
uvicorn>=0.30.6
